Move tools components to sub-dir
